+++ MSL-back/src/main/java/Maswillaeng/MSLback/controller/UserController.java
@@ -1,23 +1,14 @@
 package Maswillaeng.MSLback.controller;
 
-import Maswillaeng.MSLback.domain.entity.User;
 import Maswillaeng.MSLback.domain.repository.UserRepository;
-import Maswillaeng.MSLback.dto.user.reponse.LoginResponseDto;
-import Maswillaeng.MSLback.dto.user.reponse.UserLoginResponseDto;
-import Maswillaeng.MSLback.dto.user.request.LoginRequestDto;
-import Maswillaeng.MSLback.dto.user.request.UserJoinDto;
 import Maswillaeng.MSLback.dto.user.request.UserUpdateRequestDto;
-import Maswillaeng.MSLback.jwt.JwtTokenProvider;
 import Maswillaeng.MSLback.service.UserService;
 import Maswillaeng.MSLback.utils.auth.AuthCheck;
 import Maswillaeng.MSLback.utils.auth.UserContext;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
 @RequiredArgsConstructor
@@ -25,14 +16,6 @@ import javax.validation.Valid;
 public class UserController {
 
     private final UserService userService;
-    private final UserRepository userRepository;
-
-
-
-
-
-
-
 
 
     @AuthCheck(role = AuthCheck.Role.USER)
@@ -41,7 +24,6 @@ public class UserController {
         return ResponseEntity.ok().body(
                 userService.getUser(UserContext.userId.get()));
     }
-
 
 
     @AuthCheck(role = AuthCheck.Role.USER)
